[CBRD-20091] [MVCC] Optimize update same key (#15)
Current mvcc mechanism is changed to keep only the last version of objects in heap file and the previous ones in log. The visible version of an object will be obtained following previous versions log addresses until one that satisfies snapshot is found.

This makes retrievieng an object more specific regarding to operation type. To SELECT corresponds a basic get visible function and to UPDATE/DELETE a function to get object with lock. To get an object with lock, old versions are not relevant as we already have the last version stored in heap file, which is needed for locking.

All the implementantion of btree management regarding the same key update have been removed as current mechanism don't have to do any extra work in btree for same key update. The OID kept in btree coresponding to the heap location of an object remains the same for all its versions which means that a btree object will be updated only when the key is different.

Vacuum role is reduced because it doesn't have to clean old versions of each heap object, but only one record (the last one) in case of a delete operation and remove ins_id, prev_version_lsa when it becomes visible for all transactions.

Main modified functions:
mvcc_satisfy_snapshot():
the function had to be changed in order to diferentiate an uncommited updated record from an uncommited deleted record. Previously, in case of an uncommited update, a record had the new version inserted in heap and the previous one marked with del id and linked to the new version; the deleted version was considered visible. With current implementation, we keep only the last version which was not considered visible if it was an uncommited updated record; the non-visible return value of the function was replaced with TOO_OLD_FOR_SNAPSHOT or TOO_NEW_FOR_SNAPSHOT; the latter one is used for uncommited inserted/updated record, while TOO_OLD_FOR_SNAPSHOT is used when the snapshot is not satisfied and it's not necessary to check previous versions;

heap_get_visible_version():
this new function will return the visible version of an object, without locking it; non existent objects should be handled by the caller, some errors may be set during heap_prepare_get_record() but deleted objects should be considered of warning severity;
if the last version is TOO_NEW_FOR_SNAPSHOT, heap_get_visible_version_from_log() will be used to obtain most recent visible version stored in log;

heap_mvcc_lock_and_get_object_version():
the part of finding the visible version was removed(it is handled by heap_get_visible_version_from_log) and the entire role of the function was splitted into two major cases: get object with lock and get object without lock; getting object with lock was changed just regarding to reevaluation which is not necessary if last version is visible - evaluated already at select phase; get object without lock works similar to heap_get_visible_version but it considers non-existent-handling argument;

catcls_update_instance():
represents the old catcls_mvcc_update_instance() with an extra force_in_place argument; I thought that keeping the mvcc version would be better as it doesn't update old_attrs with new_attrs like the non-mvcc, which shouldn't be necessary if we create the new record directly from new_attrs and keep just an update_flag to tell if the update is necessary(at least one attributes is different);

heap_insert_adjust_recdes_header():
for mvcc and update operation, it reserves space for prev_version_lsa anticipatedly as this information can be obtained only at update;

heap_update_home():
mvcc operations becomes very similar with non-mvcc; the main difference is that logging was splitted into undo logging and redo logging; the final record is built only after undo logging, where the prev_version_lsa is obtained and set in header;

heap_update_relocation():
a variation of the "log_undo, update, log_redo format" from heap_update_home(), determined whether old_forward_recdes is deleted or not; if the old_forward_recdes is deleted, the redo would be just a re-delete and can be logged together with undo - in this case home record update log will use RVHF_UPDATE_NOTIFY_VACUUM;

heap_update_bigone():
for mvcc, overflow -> overflow update, the record is inserted in overflow pages and the previous version link will point to the undo-log-object of the old-home-recdes(which is a forward OID to previous-overflow-record) and the old-overflow record will be vacuumed; for remaining cases it has the same format as heap_update_home.