[CBRD-20200] Disable "in-place" update on catalog entries.
The issue is caused by not setting the link to log record with previous version when doing update-in-place.

In banana MVCC versioning design, we had a huge problem modifying schemas in certain cases. For instance, creating a class with 1000 partitions would add each partition and flush class separately. As a consequence, the instances in catalog classes were updated 1k times and for each instace we created 1k versions.

To fix the problem we used "same-transaction update" optimization. If the version we wanted to update belonged to the current transaction, instead of creating a new version we updated in-place (like a non-MVCC design would do).

Currently, we no longer have heap versions, but rather LSA links to log records where old versions can be found. From heap update point of view, the difference between non-mvcc in-place update and mvcc update is almost nothing (setting an LSA in record header is all).

One difference would be on selecting the visible version. If a class which was concurrently altered by adding 1k partition is inspected ("select * from db_class where class_name = class_with_1k_partitions"), we could indeed have to iterate through 1k log records to find the visible version. However, this is not an interesting case, as concurrent inspection with major modifications is unusual.