[CBRD-20225] Fix notify vacuum when updating REC_ASSIGN_ADDRESS; Fix checkdb issue. (#46)
We have two-fixes-in-one here:

Fix notify vacuum when updating REC_ASSIGN_ADDRESS.
After changing the MVCC versioning, the update logic was "simplified" to is_mvcc_op when we keep link to old version and !is_mvcc_op when we don't keep the link (and we always have update-in-place). The problem is that vacuum notification is also linked to "is_mvcc_op". When we update a REC_ASSIGN_ADDRESS (as we do with catalog entries), we do not keep link to old version but we need to notify vacuum.

Fix checkdb issue.
I fixed it here, because the second issue reproduced once first was fixed. This is same as CBRD-20234 issue.
The problem was one change we made earlier to locator_check_btree_entries. We copy heap records instead of peeking. However, the copy algorithm chooses to copy to record data directly if it has a non-null value or to scan_cache->area if the record data was NULL.
Our record data was not initialized... actually nothing in record was initialized. The record area_size was smaller than record size and the result of heap_next was S_DOESNT_FIT. The code only counts the number of successful returns of heap_next.
Fixed by initializing the record used to copy from heap.