[CBRD-20158] String compression (#160)
http://jira.cubrid.org/browse/CBRD-20158

Strings are now saved on the disk in a compressed form, using the same compression function as the one used in the log zip one. This is applied, for now, to VARCHAR and VARNCHAR types. The compression is done before putting the strings in the OR_BUF buffer, more precisely in the or_put_varchar_internal function.

The current encoding for the VARCHAR and VARNCHAR types with size larger than 255 (0xFF) : 
[0xFF][4_bytes_compressed_data_size][4_bytes_decompressed_data_size][compressed_data].

If the string is supposed to be compressed, however its compressed_size > original_size - 8, then we set compressed_size = -1, and add to the buffer decompressed_size of the original data.

There is one regression that needs revision, and that is: sql/_23_apricot_qa/_01_sql_extension3/_02_new_sql_types/_01_enum/cases/_18_adhoc_multiple.sql. It fails because the size of the enum string exceeds the maximum disk size of such a string. However now, with the compression enabled, it will not fail anymore since the size will be greatly reduced by the compression.