[CBRD-20421] Fixed bad set for compressed_length (#179)
http://jira.cubrid.org/browse/CBRD-20421

In case of a failed compression due to compressed_length >= original_size - 8, compressed_length was set to -1. But compressed_length is an unsigned value so in this case everything would go wrong. Changed the compressed_length to be set to 0, in case of such event.

It should also fix the memory leak issue since I think that would come from the assert that caught the error. Also I have switched the asserts to catch those that might be set bad.