[CBRD-20421] Removed _setjump in heap_attrinfo_transform_to_disk_internal (#185)
http://jira.cubrid.org/browse/CBRD-20421

Due to bad error handling of _setjmp, in or_put_varchar_internal, if an error like ER_TF_BUFFER_OVERFLOW was encountered, it would jump out of the function and let string and wrkmem leaked. We have removed the _setjmp in this scenario and handle the errors that might appear.

We might also want to remove any _setjmp usage other than this, to have better error handling in these situations.