[CBRD-20340] Redesign the log page buffer. (#184)
http://jira.cubrid.org/browse/CBRD-20340
The log page buffer was redesigned to be latch-free.

The hash table which was used for the log page buffer was replaced with an array, and the mutex used for the log buffer access was deleted. Before, there was a critical section and  a mutex,  and now the mutex was removed and replaced with some assertions to be sure that we are in the critical section. This way, the safety is kept. 
The log_buffer was redesigned - the fields fcnt, ipool, recently_freed, flush_running, dummy_for_align were removed, the log_bufarea structure was removed, and also log_pb_global_data was redesigned and the allocation of the log pages. Now, the logpages are allocated into the pages_area from the global log_Pb, the log buffers are allocated into the buffers field from the log_Pb, and every log page buffer has a pointer to a log page. 

The function logpb_fix_page was redesigned too, to keep the new implementation, and was renamed to logpb_locate_page. This function returns the corresponding log page from the log buffer, if it exists, or it brings it from the disk. The function logpb_replace was removed, because it was no longer needed in the current implementation.

The functions related to fcnt , like logpb_unfix_page, logpb_free_page, logpb_free_without_mutex  were removed. Same from the functions related to the old log page hash, like logpb_print_hash_entry.

The logpb_copy_page function was also modified to keep the new implementation. Another minor modifications involve that some return codes have changed, for example from LOG_PAGE* to int, because the log page was already returned in one of the parameters, and the return code was set to the error code.

The delayed_free_log_pgptr variable, used for safety flushing, was also removed, because in this version it is not guaranteed that the pointer is kept in the buffer, and it's functionality was replaced with some other checks, using log_Gl.append.prev_lsa and log_Gl.hdr.append_lsa. Another statistics were deleted, like log_num_fetch_ioreads, last_delayed_pageid and total_delayed_page_count.