[CBRD-20471] fix overwriting log_page_buffer record with end of log record during flush (#215)
The log page buffer was replaced with an end of log when flushing, for safety reasons. Other threads could read wrong the log page buffer, so the end of log should be appended to a copy, not to the buffer itself. 

Moreover, the page should have the dirty flag cleared if it is in the buffer, because it could be flushed more than once during one flush.