changed condition in logpb_end_append (#219)
The old condition for setting the dirty flag in logpb_end_append, that the log_Gl.append.prev_lsa.pageid and log_Gl.hdr.append_lsa.pageid must be equal was wrong. The condition should be that the append.prev_lsa and hdr.append_lsa must be different to set the dirty flag.