[CBRD-20504] Update fix count check (#241)
http://jira.cubrid.org/browse/CBRD-20504

The fix count check was outdated. A while ago we changed access on vacuum data pages and how they get flushed on checkpoint. In the crash, checkpoint thread also fixed page during fix count check. Fix count was 2 instead of 1.

The purpose of the check was not to leak any pages. Since this is vacuum data thread, the check now has two components:
1 verify first and last vacuum data pages are only fixed once.
2 get number of page holds. it should be one if last page and first page are the same, or two otherwise.