[CBRD-20162] MVCC record header changed to always keep CHN (#235)
The mvcc record header has a union for the chn and the delete id, and we want to keep always the chn, so the union was removed and were created fields for chn and delete id. This implies that we have to change the way we access the chn. Also, we could have the delete id and the insert id in parallel, and we have to count the chn field always, when we want to restore the data. The chn logging (used when chn and delete id couldn't be together) was removed.

The long_chn flag was also removed and when it was needed it was replaced with a null delete id flag, and there were added some checks to see if it is a null delete id flag.