[CBRD-20447] Implement a function that returns the size of a value on disk (#270)
We need a new function in the database to show us the size of a value on disk. This is useful because we may have compressed strings and we want to see if the compression works and how much it helps.
The new function implies that e have to modify the grammar and the lexer, to add a new rule for "disk_size". The function calls pr_data_writeval_disk_size for a db_value, which calculates exactly the size on disk.