fix system ops
I had some issues with "topope commit with postpone". The first issue is that we needed to extend it for all new system operation end types (undo, compensate, run postpone). If we were to use the previous approach (new log record types for each end system op), we needed three new log record types, with three new log record structures and handling them everywhere (analysis, redo, undo, rollback, finish postpone, dump so on). The implementation was starting to get ugly and unreasonable.

I chose to sacrifice log size for an easier approach code-wise. I keep only one log record for end system operation, that includes the end system operation type (commit, abort, logical undo, logical compensate and logical run postpone). Each end type has a special functionality in some contexts while acting like a committed system operation in most contexts. Therefore, it is easier to check in each context the end type and handle it accordingly. We avoid duplicating a lot of code and if we will add new end types in the future, it will be easier to do it with current approach.

System op start postpone now includes the log record for end postpone. If the server crash during system operation postpone, we now know how to end it (based on that log record).

I also missed a case when I first wrote the extention to system operation: logical run postpones are needed during a system operation postpone. Postpones are used for:
1. dropping file - logical run postpone system operation, during transaction postpone
2. unreserving file sectors - physical run postpone, during system operation postpone, during transaction postpone (called by #1)
3. deallocating pages (from file header) - logical run postpone system operation, during system operation postpone
4. notifying vacuum - logical run postpone without system operation, during transaction postpone
5. removing index statistics - logical run postpone without system operation, during transaction postpone
6. serial index mark delete - logical run postpone with/without system operation, during transaction postpone
7. changing delete new home from non-reusable slot to reusable slot - physical run postpone, during transaction

The difficult cases are 1, 2, and 3. 1 and 2 mean that we need to support postpones in a logical run postpone system operation. 3 means that we need to support logical run postpone during a system operation postpone. Which from recovery point of view are very complicated cases.

First of all, if we meet the end of a logical run postpone in transaction state TRAN_UNACTIVE_TOPOPE_COMMITTED_WITH_POSTPONE, there are two possible scenarios:
1. this is the postpone of logical run system operation
2. this is the end of logical run system operation started during the postpone of a system operation.
The outcome should be different in the two cases. To handle them properly, we include is_sysop_postpone information in end run postpone system operation. If this is true during recovery, then we must be in the second scenario. The state TRAN_UNACTIVE_TOPOPE_COMMITTED_WITH_POSTPONE is preserved and postpone LSA of system operation is updated. If it is false, then it must be first scenario, the transaction state becomes TRAN_UNACTIVE_COMMITTED_WITH_POSTPONE and the postpone LSA of transaction is updated.

Another complication is that we can have a crash during the logical run postpone system operation during the postpone of another system operation. Which means we need to rollback it first before recovering the postpone of system operation. To detect the case, we save the LSA of system op start postpone. If undo_nxlsa is greater than this, then we need to do a partial rollback.

As a consequence to all these changes, saving system operations that started postpone was also updated.

I really need to write a TDD for these changes.