[CBRD-20448] New system parameter use_compression (#280)
I added a new parameter named PRM_ID_ENABLE_STRING_COMPRESSION, which turns on and off the compression for the string. The string format remains the same on disk (under 255 it contains one byte with the length and then the string itself, after 255 it contains a byte with the value 255 ,then the compressed length, the decompressed length, the string and the alignment), but the compression length is set to 0 if we set the parameter to false, meaning that there is no compression and the string is stored in its original size.