[CBRD-20497] Performance enhance for string compression (#246)
http://jira.cubrid.org/browse/CBRD-20497

This is a work in progress but I would like to open the PR for easier discussion.

In order to avoid multiple compresses of the same DB_VALUE, we will now be storing the string compression as well as its length, until the clear of the DB_VALUE. Therefore every time we will need the compression of the string we might be able to take it from the DB_VALUE, since it may have already been computed in a lengthval type function.

I have added 3 more fields in the DB_CHAR structure:

compressed_string : Field in which we will store the compressed string of the current DB_VALUE.
compressed_length : Field in which we store the length of such a compressed string, if it exists.
compressed_need_clear : Field that states whether the compressed string needs a clear or not.
Now every time we would need for the compression of the DB_VALUE we will be checking whether it has been already computed or not. And if it has not been computed so far, we do it on spot and then continue with the process.

The compressed_string and compressed_size fields work like this :

If compressed_size is set to 0 then the DB_VALUE has not been through a compression process before, so whoever checks the state of this field will also do the compression on spot since it will be needed at some later point.

If compressed_size is set to -1 and size is lower than PRIM_MINIMUM_STRING_LENGTH_FOR_COMPRESSION then the string stored in the DB_VALUE does not require a compression due to its size.

If compressed_size is set to -1 and size is greater than PRIM_MINIMUM_STRING_LENGTH_FOR_COMPRESSION then the DB_VALUE has been prompted to a compression before, but due to its bad compression it failed so another step of compression is not needed.

If compressed_size is greater than 0, then the DB_VALUE has stored in it the compressed_string as well and can make use of it whenever it needs.