[CBRD-20379] New cleanup criteria for XASL cache (#289)
http://jira.cubrid.org/browse/CBRD-20379
XASL Cache was cleaned only when the maximum capacity is reached. I introduced another criteria, based on time. After a certain amount of time, the oldest entries are cleaned. I introduced a new system parameter for the time threshold and made a new cleanup type. For this type, it is used an array for storing the entries instead of a binary heap.