[CBRD-20572] Fixed leak of bad value coercion
http://jira.cubrid.org/browse/CBRD-20572

Since VARCHAR and VARNCHAR DB_VALUE types will now hold a possible compressed_string, when coercion is done, in some cases it would just change the type of the original DB_VALUE, therefore it might have become a CHAR or NCHAR type, types that could not have a compressed_string field in it. So now we just have to copy in the destination instead of just switching the types.