[CBRD-20511]  Add code to detect leaked page fixes with ordered fix (#295)
http://jira.cubrid.org/browse/CBRD-20511
I added some checks to be sure of the integrity of the page watchers. Since any page watcher contains a magic number which will be modified eventualy in case of the corruption, I checked at pgbuf_fix and pgbuf_unfix that the page watchers have this value unmodified. This checks are made only in debug mode.