[CBRD-20690] btree_rv_record_modify_internal: remove outdated assert (#357)
http://jira.cubrid.org/browse/CBRD-20690

Removed the assert that checked bug info flag is not set. Logging new key insert, when the key is overflowing, is now done under a system operation and is logged with RVBT_RECORD_MODIFY_UNDOREDO. The debug info is not really saved to undo data (there is no undo data), but is saved in redo data, so the flag is set in rcv.offset).

Therefore, the assert is no longer valid.