[CBRD-20703] file numerable find nth: cache last search location (#365)
http://jira.cubrid.org/browse/CBRD-20703

This patch purpose is to optimize index loading. Specifically it tries to optimize file_numerable_find_nth function which takes a lot of loading index time because it goes through many of user page table pages every time is called.

Since the pattern of file_numerable_find_nth is to find nth, then nth+1, then nth+2 and so on, we can cache the location in user page table for last search. Next search is likely find the page it's looking for in the same user page table page or in next, significantly reducing the number of fixed/unfixed pages.

Note that deallocating a page will reset the cached search location (external sort never calls it anyway, so it is just a safeguard).