[CBRD-20702] improve file extensible data merge (#371)
http://jira.cubrid.org/browse/CBRD-20702

The issue is caused by file header inconsistency. When there is no full sector, we expect that the full sector table has only one empty component. In the core file we have to empty components due to the way extensible data page were merged: always trying to merge current page (where data is removed) with next page.

Changed the merging routine of extensible data to check pages merges with both previous and next pages. At the same time, I unified some repeated code in a single function.

Modified the recovery of extensible data merging to make it simpler (and be independent of the various cases). Even though we log more data, this is a rare event that will not have a noticeable impact on overall log size.