[CBRD-20881] Bind CTE names and types separately, prior to the query binding (#467)
http://jira.cubrid.org/browse/CBRD-20881

The issue was caused by the name resolving of the CTEs. The scope used to resolve a CTE contained higher level specs, which was incorrect - each CTE should be resolved independently, with access only to the entities it contains.

The names and types of a CTE were resolved for each of its occurences. It required to alternatively set the non-recursive/recursive parts to NULL in order to handle self references, which was difficult to follow and control.

The fix:

pt_bind_names_in_cte has been created to handle the binding of a CTE. In this way the CTEs are resolved only once, prior to the query and other CTEs resolving - where it may be referenced.

pt_bind_types was renamed to pt_bind_spec_attrs and its functionality has been split in pt_get_attr_list_of_derived_table and pt_set_attr_list_types. These functions were needed in pt_bind_names_in_cte also.