[CBRD-20932] logtb_is_interrupted_tdes: allow vacuum interrupt during recovery (#502)
http://jira.cubrid.org/browse/CBRD-20932

Loosen-up safe-guard. During recovery, we can allow interrupts even if thread is vacuum worker.

Note that the thread is not actually a vacuum worker. It is hacked into thinking it is a vacuum worker while recovering the work of the vacuum worker. This is beyond the scope of initial safe-guard, which only tries to ensure a good control on errors during vacuum job execution.