[CBRD-20669] heap_remove_page_on_vacuum: accept the case with checkpoint thread waiting for latch (#500)
http://jira.cubrid.org/browse/CBRD-20669

The crashed safe-guard is there for a clear purpose: we tried to make sure no worker will try to access a heap page deallocated by vacuum. While this is true for all active and vacuum workers, the checkpoint thread cannot be prevented from try to latch page (and we don't to).

Change safe-guard check from is any waiter on page to is any non-checkpoint waiter on page.