[CBRD-21081] Removed assert regarding the match of a BTREE operation (#584)
http://jira.cubrid.org/browse/CBRD-21081

In the scenario provided, it happened to rename a table, having an auto-increment key, to a new name, then back to its original name. In this case, one key would be marked for insert, while the other would be marked for insert and delete, so therefore the MVCC info would not match, eventually failing an assert. We remove the assert since it was incorrect.