[CBRD-21352] log_final: reset log_GL.rcv_phase (#720)
http://jira.cubrid.org/browse/CBRD-21352

The issue is caused by shutdown/restart server in the same execution instance. Global variables are not always properly reset.

This case was caused by log_GL.rcv_phase. We supress **pgbuf_is_temporary_volume** check when server is not yet restarted. This may be called before disk cache is loaded, which is problematic and a safe-guard is hit.

In our case, restoreslave first calls **boot_restart_from_backup**, shuts it down and then **db_restart** to apply HA logs. Since log_GL.rcv_phase was not properly reset during shutdown, when volumes are booted, it crashes calling pgbuf_is_temporary_volume (debug mode).

This fixes log_GL.rcv_phase, which proved to be enough for the case described in issue. However, I doubt this is the only issue we face because globals are not properly reset.