[CBRD-21367] logpb_checkpoint, logpb_backup: sync with disk extension
http://jira.cubrid.org/browse/CBRD-21367

fileio_find_next_perm_volume is not synchronized with disk extensions in any way. so, it may end up accessing a volume not yet fully initialized, which can ends up hitting safe-guards. logpb_checkpoint and logpb_backup are the two cases that call it in a running server and may face issues.

Faced by locking disk extension while calling the function (or deciding what the last volume is).