[CBRD-21415] fix recovery for file_tracker_unregister crash (#750)
http://jira.cubrid.org/browse/CBRD-21415

This is a complicated case.

1. file_destroy (during postpone)
2. file_tracker_unregister (file_dealloc).
3. sysop with logical undo starts postpone
4. crash!

We have a sysop postpone phase inside a logical undo, inside a logical run during transaction postpone phase. The recovery code was not adept to handle the case, so it had a safe-guard to ban the case.

This is a hack/quick fix that should work in this case. We allow a sysop start postpone in this context and try to correct the transaction state changes. However, it may not work in other complicated cases (if they exist or if they will be added). A correct way to manage nested system operations is log more information (e.g. previous transaction state).