[CBRD-21483] disk_rv_undo_format reset volume purpose (#784)
http://jira.cubrid.org/browse/CBRD-21483

it's an outdated safe-guard. before adding volume, because page buffer must know which volumes are permanent and which are temporary, the number of permanent volumes is incremented and volume purpose is set.

previous disk_rv_undo_format fix now redirected case handling and hit the assert that checked purpose is unknown.

fix by removing safe guard and resetting purpose.