[CBRD-21433] Enhancement for loading a foreign key (#786)
http://jira.cubrid.org/browse/CBRD-21433

The loading of a foreign key will have the values checked similar to how the merge phase of merge-sort works. Instead of checking each value from the foreign key to reside in the referenced column of the primary key by starting a new traversal search in the b-tree of the primary key, will now use only 1 traversal search for the first value. The next values will be checked by starting from the position of the last value found in the primary key, instead of starting another traversal search on the b-tree.

This proves to be an improvement as starting new searches in large primary keys would lead to inefficiency during the loading of the foreign key. The new check is done just after the build of the leafs of the foreign key and leading to success will continue to build the foreign key unless any other errors appear.