[CBRD-21583] vacuum_process_log_block: sa_mode_partial_block was interrupted (#834)
http://jira.cubrid.org/browse/CBRD-21583

This seems to be just a safe-guard issue. We usually mark in vacuum data when jobs are started and when they are finished, so we can adjust safe-guards after restart (expecting already vacuumed heap pages).

However, stand-alone mode vacuuming partial blocks can be interrupted and will not be flagged in any way. Therefore, it is better to consider that job was interrupted when sa_mode_partial_block is true.