[CBRD-21617] vacuum_heap_page: fix interrupted case (#847)
http://jira.cubrid.org/browse/CBRD-21617

Fixed code that tries to handle interrupted vacuum jobs and deallocated heap pages. One of the cases it used to consider a heap page was deallocated is when the number of slots is less that what it expected.

However, that is not always true. Another possible case is undoing a series of inserts. The number of slots are incremented while inserting (and vacuum is logged for each insert), but on rollback the number of slots are decremented (this doesn't always happen, only when the deleted slot happens to be last).

The crash was caused by a safe-guard that was supposed to be true in case page was really deleted. Since the page was not deleted, the safe-guard failed.

Heap pages are no longer assumed deleted and usual checks are executed on each record.