[CBRD-21579] alternative fix: generate one eval structure for an analytic function when optimization fails (#856)
* [CBRD-21579] Fix analytics optimization with >= 4 expressions

http://jira.cubrid.org/browse/CBRD-21579

Fixed by @rimio

Remove eval created on non-leaf level and order children by level.

To be tested thoroughly with 4 or more analytic expressions.

* [CBRD-21579] alternative fix: generate one eval structure for an analytic function when optimization fails (#838)

http://jira.cubrid.org/browse/CBRD-21579

This is an alternative quick fix for the cases.
When an invalid evaluation structure is generated, we regard optimization is failed and generate trivial list.
I believe this fixes all the regression of b72ce30. However we should tackle issues of pt_optimize_analytic_list.