[CBRD-21814] Receiving error from server may not be completed (#951)
http://jira.cubrid.org/browse/CBRD-21814

This is a quick fix to the issue.

To disconnect the temporary connection for the cases right after sending error to client prevents the client from properly receiving the entire error area. Client may access bad memory and crashes.
It is a legacy issue and we need a complete fix.