[CBRD-21674] - Added support for json union (#961)
http://jira.cubrid.org/browse/CBRD-21674

We need a schema in order to validate a json when inserting a new value in table. In order to do that we have the JSON_VALIDATOR object.

We will store in cache a tp_domain pointer which will be used further for the retrieval of the domain.

When trying to unify two types ("/src/query/list_file.c", method: qfile_unify_types) we will check if the domain pointers are the same, but when we are dealing with json types, if the schemas are not the same, those pointer will be different so we need an additional check to pass that condition and to successfully make the union.