[CBRD-21672] - Support SQL/JSON path and other functions (#928)
This PR includes the following functionalities:
-> JSON_ARRAY_APPEND
-> JSON_SET
-> JSON_REPLACE
-> JSON_KEYS
-> JSON_GET_ALL_PATHS
-> Conversion from SQL to JSON path and viceversa

The conversion from JSON to SQL path is implemented in function: db_json_convert_rapidjsonpath_to_mysqlpath(char raw_path, const JSON_DOC &doc, std::string &converted_path)
In the converted_path variable we will have the transformation.

The logic is simple: first we need to erase trailing and leading spaces and check if the path is already converted. We iterate through the path and get the tokens between "/" separator in only one traversal. We use the current token and check if we encounter a number or a string to verify it is an index or a name and append it to the result. For special characters we use a hashmap to store the corresponding transformation.

The conversion from SQL to JSON is pretty similar, just we have more separators .[]"

The JSON_GET_ALL_PATHS function requires one parameter which is a JSON_OBJECT and returns an array containing the SQL path to each object in that json. Recursively we go until we reach each final object and construct the path on the way.
For example suppose we have the next json_object: '["a", "b", {"a": 1, "b": 2, "c": {"d": 4}}]'

The next functions should have the same behaviour like the MYSQL standard:

-> JSON_ARRAY_APPEND: https://dev.mysql.com/doc/refman/5.7/en/json-modification-functions.html#function_json-array-append
-> JSON_SET: https://dev.mysql.com/doc/refman/5.7/en/json-modification-functions.html#function_json-set
-> JSON_REPLACE: https://dev.mysql.com/doc/refman/5.7/en/json-modification-functions.html#function_json-replace
-> JSON_KEYS: https://dev.mysql.com/doc/refman/5.7/en/json-search-functions.html#function_json-keys