[CBRD-21871] to keep inserted object template for getGeneratedKeys (#987)
http://jira.cubrid.org/browse/CBRD-21871

It is a regression of #984
Before #984, it accessed freed object template. It worked as a hack.

To access inserted object template, client insert routines request to template manager to keep it by resetting discard_on_finish. Since the flag was not used, fix changes its behavior. Caller who wants to keep the template after flush should explicitly free it.