[CBRD-21875] json: increment index correctly when parsing sql path (#995)
http://jira.cubrid.org/browse/CBRD-21875

Expected json path will be parsed correctly

When parsing the sql path, if we have an array index like $.[123] we need to move the index of the iteration i to the index of the last ending bracket. This way the index will be auto incremented in the for loop.

This PR should fix the parsing of numeric index in sql path.