[CBRD-21887] Wrong result sets were returned by json_set (#996)
http://jira.cubrid.org/browse/CBRD-21887

Wrong result sets were returned by json_set

Right now, json_insert and json_set will call the same method for the insertion logic, because they should be consistent. 

We can insert a new value at a specified path if the type of that node is JSON_OBJECT or JSON_ARRAY, otherwise we will throw an error.

We can not insert specifying an index in a JSON_OBJECT or specifying a key in a JSON_ARRAY. This will throw an error.

If we have a JSON_ARRAY and insert to a position greater than the length of the array it will fill with null values until we reach the desired position.