[CBRD-21889] Handle duplicate keys in json_keys (#993)
http://jira.cubrid.org/browse/CBRD-21889

Handle duplicate keys in json_keys
This PR resolves also:
http://jira.cubrid.org/browse/CBRD-21912
http://jira.cubrid.org/browse/CBRD-21770

First of all, added also some changes to existing code in order to avoid duplicate code.

We reached to the conclusion that we do not allow duplicate keys in a json, so if we meet this scenario, we will throw an error.

Created a JSON_BASE_WALKER class, to be easiliy extended for future traversal implementations of the json "tree" structure. It should be extended and implement the methods CallBefore or CallAfter. This aproach traverse the tree recursively. 

Using JSON_DUPLICATE_KEYS_CHECKER class, we check for duplicate keys at parse time and throw an error if two keys are the same, only on the same level!