[CBRD-21745] Fix db_make_string and db_make_db_char signature (#1020)
http://jira.cubrid.org/browse/CBRD-21745

Since the DB_VALUE will store strings as char * and not as const char *, we encountered some issues and ambiguities regarding to the functions we used to store those strings, such as db_make_string and db_make_db_char. These two used, as parameters, const strings, which is not entirely correct. In order to keep the const property when needed, we have decided to use db_make_string when we want to keep and store the pointer referenced as parameter, while db_make_string_copy will, instead, create a copy of the referenced pointer. I have adjusted the calls throughout the code to reflect these changes.