[CBRD-21980] [Regression] [HA] tables can't be replicated to slave node (#1066)
http://jira.cubrid.org/browse/CBRD-21980

The resume_status member variable of a thread context wasn't reset when recycling or retiring a context, so the if (thread_p->resume_status == THREAD_RESUME_DUE_TO_INTERRUPT) at log_writer.c:xlogwr_get_log_pages would execute and would make the entry's eof_lsa not null. The whole logic in xlogwr_get_log_pages relies on the fact that entry's eof_lsa is null, so this was generating a replication bug.