[CBRD-22123] increase css_Maximum_server_count to 1K (#1115)
http://jira.cubrid.org/browse/CBRD-22123

This bug appears because cub_master gets blocked when delegating the connection to cub_server because the socket send is blocking and cub_server spawns receiving threads instead of pooling them, which has overhead. Because cub_master has a backlog of only 50, clients quickly saturate this value and they get rejected.

We decided, for the time being, to simply increase the cub_master backlog to 1K in order to fix the regression. We conveyed that pooling threads on the cub_server side it's not worth it and we may, in the future, separate connections to threads on cub_master's side.