[CBRD-22153] lock demotion primitives (#1121)
http://jira.cubrid.org/browse/CBRD-22153

lock_demote_class_lock is the interface function for server-side lock manager and locator_demote_class_lock is the client api.

We don't have usages of these functions yet. It will be used for online index loading task.

We may need a lock promotion api for client to restore ex-lock. This might be done as a separated issue if needed.