[CBRD-22232] [JSON]Cannot coerce value of domain "national character" to domain "json" (#1168)
http://jira.cubrid.org/browse/CBRD-22232

The implementation of db_json_get_json_from_str does doc.Parse(str). Because the string is not null terminated, it will go until reaches '\0' so this is why we will have a parse error.

Modified the header of db_get_json_from_str to send a string with a specific length.